KAN-56

+++ app/src/controllers/postController.js
@@ -12,7 +12,7 @@ async function handleGetPost(postID){
         let json = await serverRequest.makeRequest(`/posts/${postID}`, requestOptions);
 
         if (json.error) {
-            return {"error": json.error}
+            return {"error": json.error, "status": json.status};
         }
 
         let post = json["post"];
@@ -22,12 +22,18 @@ async function handleGetPost(postID){
         return post;
     } catch {
         console.log("Error getting post.");
-        return {error: "Error getting post."}
+        return {error: "Error getting post.", "status": 500};
     }
 }
 
-async function handleDeletePost(postID, jwt){
+async function handleDeletePost(options){
     try {
+        if (!options.jwt || !options.username) {
+            return {error: "You must be logged in to delete a post.", "status": 401};
+        }
+
+        let {postID, jwt} = options;
+
         const requestOptions = {
             method: 'DELETE',
             headers: {
@@ -39,20 +45,26 @@ async function handleDeletePost(postID, jwt){
         let json = await serverRequest.makeRequest(`/posts/${postID}`, requestOptions);
 
         if (json.error) {
-            return {"error": json.error};
+            return {"error": json.error, "status": json.status};
         }
 
-        return { success: true };
+        return { success: true, "status": 200 };
 
     } catch(error) {
         console.log("Error deleting post.", error);
-        return {error: "Error deleting post."};
+        return {error: "Error deleting post.", "status": 500};
     }
 }
 
 
-async function handleWritePost(boardName, title, content, username, jwt){
+async function handleWritePost(title, content, options){
     try {
+        if (!options.jwt || !options.username) {
+            return {error: "You must be logged in to post.", "status": 401};
+        }
+
+        let {boardName, username, jwt} = options;
+
         const requestOptions = {
             method: 'POST',
             headers: {
@@ -64,22 +76,30 @@ async function handleWritePost(boardName, title, content, username, jwt){
 
         let json = await serverRequest.makeRequest(`/posts`, requestOptions);
 
+        console.log("status from handleWritePost: ", json.status)
+
         if (json.error) {
-            return {"error": json.error};
+            return {"error": json.error, "status": json.status};
         }
 
         console.log("post successfully made: ", boardName);
-        return { success: true };
+        return { success: true, "status": 200 };
 
     } catch(error) {
         console.log("Error writing post.", error);
-        return {error: "Error writing post."};
+        return {error: "Error writing post.", status: 500};
     }
 }
 
 
-async function handleWriteComment(postID, comment, username, jwt){
+async function handleWriteComment(options){
     try {
+        if (!options.jwt || !options.username || !options.comment || !options.postID) {
+            return {error: "You must be logged in to comment.", "status": 401};
+        }
+
+        let {username, jwt, comment, postID} = options;
+
         const requestOptions = {
             method: 'POST',
             headers: {
@@ -94,18 +114,24 @@ async function handleWriteComment(postID, comment, username, jwt){
         console.log("server response from handleWriteComment: ", json);
 
         if (json.error) {
-            return {"error": json.error};
+            return {"error": json.error, "status": json.status};
         }
 
-        return { success: true };
+        return { success: true , "status": 200};
 
     } catch(error) {
         console.log("Error writing comment.", error);
-        return {error: "Error writing comment."};
+        return {error: "Error writing comment.", "status": 500};
     }
 }
-async function handleEditPost(postID, title, content, jwt){
+async function handleEditPost(postID, title, content, options){
     try {
+        if (!options.jwt || !options.username) {
+            return {error: "You must be logged in to edit a post.", "status": 401};
+        }
+
+        let {jwt} = options;
+
         const requestOptions = {
             method: 'PATCH',
             headers: {
@@ -117,14 +143,14 @@ async function handleEditPost(postID, title, content, jwt){
 
         let json = await serverRequest.makeRequest(`/posts/${postID}`, requestOptions);
         if (json.error) {
-            return {"error": json.error, "post": json.post};
+            return {"error": json.error, "post": json.post, "status": json.status};
         }
 
         return { success: true };
 
     } catch(error) {
         console.log("Error editing post.", error);
-        return {error: "Error editing post."};
+        return {error: "Error editing post.", "status": 500};
     }
 }
